requirements.txt
mysqlclient==2.1.1
sqlalchemy==1.4.41
python-dotenv==0.21.0